Disable workspaces (having issues)
Some buffers seem to not get added to the workspace, which seems to me
to be a bug. I made a bug for that:

https://github.com/hlissner/doom-emacs/issues/3468

Also, from the documentation of the workspaces module, there's some
behaviors I don't really want around automatic workspaces. So I might
need to revisit, to customize those behaviors, but I don't really have
time for it at the moment.